notification/requirements.txt
black==22.10.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.2
pika==1.3.1
platformdirs==2.5.4
python-dotenv==0.21.0
